requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
awesomeversion==25.8.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.10.6
cryptography==46.0.1
deepmerge==2.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.4
packaging==25.0
passlib==1.7.4
playwright==1.54.0
pluggy==1.6.0
ply==3.11
propcache==0.3.2
pyasn1==0.6.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyipp==0.17.2
PyJWT==2.10.1
pysmi==1.4.4
pysmi-lextudio==1.4.3
pysnmp==6.2.6
pysnmpcrypto==0.0.4
pytest==8.4.2
pytest-cov==4.1.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-nmap==0.7.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zeroconf==0.147.0
